+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+
+- Add screenshotId to screenshot file name.
+- Delete existing screenshots before saving new screenshots.
+
 ## 0.1.0
 
 - Remove indexToScreenshot from config.